Fixes shadowed variable warning, under ruby 2.3.0
This patch addresses a number of `warning: shadowing outer local variable - key`
issues that cropped up when using this gem (via a `thor` dependency)
with ruby 2.3.0.

There are some other warnings that are also happening under this version
of ruby that this patch doesn't address, as those seem like a bit more
involved and should probably be handled by someone more familiar with
this codebase.